Cleanup, verifier updates
Fixed some more verifiers.
Cleaned up the bot and removed unused/dead sites
