+++ DiceBot/NitrogenSports.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Security.Cryptography;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -76,7 +77,53 @@ namespace DiceBot
                 catch { }
             }
         }
+        new public static decimal sGetLucky(string server, string client, long nonce)
+        {
+            HMACSHA512 betgenerator = new HMACSHA512();
 
+            int charstouse = 5;
+            List<byte> serverb = new List<byte>();
+
+            for (int i = 0; i < server.Length; i++)
+            {
+                serverb.Add(Convert.ToByte(server[i]));
+            }
+
+            betgenerator.Key = serverb.ToArray();
+
+            List<byte> buffer = new List<byte>();
+            string msg = client + "-" + nonce.ToString();
+            foreach (char c in msg)
+            {
+                buffer.Add(Convert.ToByte(c));
+            }
+
+            byte[] hash = betgenerator.ComputeHash(buffer.ToArray());
+
+            StringBuilder hex = new StringBuilder(hash.Length * 2);
+            foreach (byte b in hash)
+                hex.AppendFormat("{0:x2}", b);
+
+
+            for (int i = 0; i < hex.Length; i += charstouse)
+            {
+
+                string s = hex.ToString().Substring(i, charstouse);
+
+                decimal lucky = int.Parse(s, System.Globalization.NumberStyles.HexNumber);
+                if (lucky < 1000000)
+                {
+                    lucky %= 10000;
+                    return lucky / 100;
+
+                }
+            }
+            return 0;
+        }
+        public override decimal GetLucky(string server, string client, int nonce)
+        {
+            return sGetLucky(server, client, nonce);
+        }
         void GetStats()
         {
             Thread.Sleep(1);
